refactor: use phantom type parameters rather than new types
We collapse all the sorts of "global" names into one type, indexed with
a phantom parameter tagging what sort they are. We keep the local names
separate even though they could be collapsed with the globals -- this is
because we will shortly add module scoping to the globals (but not the
locals).