refactor: disambiguate local term/type variables
We clarify which local variable names are (expected to be) term
variables and which type variables, via a phantom parameter.

Due to the two types of variables being in the same namespace and thus
shadowing each other, we unfortunately have to ignore the difference
rather often, negating some of the benefits of having the stricter
typing.

(This is a big commit, as it is one logical unit. It is at least rather
mechanical.)