refactor: Typechecker local context lookups
We expose separate functions for looking up a type variable and a term
variable by name. They return the appropriate error if the variable
names something in the context of the wrong sort, or if the variable is
not present in the context. This is in preparation for splitting
'LVarName' into two types: one for term variable names and one for type
variable names.