refactor: convert varRefName to a function
This was only ever used as a standalone getter. Since it does not make
much sense to change the Name inside a VarRef without caring whether it
is a local or global, we remove the ability to set via the lens. This
could be done by writing a Getter, but there does not seem any advantage
in using that over a plain function.