feat!: add module names
This is a large commit, but most of it is pretty mechanical.

NB: we remove the test 'unit_rename_def_capture'. It tested that we
could not rename a definition in such a way that a reference would be
captured by a local binding. However, now that global variables have
module-qualified names, this is no longer possible: a local binding is
not considered to shadow any globals.

BREAKING CHANGE: this change requires a database migration, as it
changes the representation of `Prog`. However, since this is just
serialised to json and stored as a blob in the DB, it requires no schema
changes. Since we have no programs we need to preserve, we decided not
to bother with a migration. This means that DBs created before this
commit will not load with a primer containing this commit.