refactor: use modules where possible
Instead of using 'defaultTypeDefs', define and use 'builtinModule' and
'primitiveModule'. This is in preparation for adding module names and
taking modules more seriously.

There is one theoretical drawback: in Tests.EvalFull we lose the ability
to generate definitions for global variables. However, we never actually
used this ability, so there is no loss in practice. (However, we may
want to re-instate this ability in the future so we can generate an
arbitrary context in these tests.)