feat: action to rename the editable module
The reason for the slightly convoluted implementation is that I am
looking forward to when we have multiple editable modules, where we need
to edit multiple modules (as they all may contain references to the
renamed module). A potential optimisation is only checking for name
clashes with imports and not rewriting theose modules, since we do not
support renaming imports, and imports cannot depend on the editable
module.

I also take the chance to simplify the test code which does module
renaming, now that the idea has been properly implemented in the
library, with name clash detection. It is not worth making the library
more complex or exposing an "unsafe" version for one use in testsuite
(which intentionally clashes names, in order to merge modules).