refactor: Add a type alias for `Map GVarName Def`.
We're using this type all over the place, so we might as well give it
a convenient name.

Note that `Primer.Eval` had an alias for this type called `Global`,
but I've replaced that here with `DefMap`, because while the type is
used in eval to represent global variables (i.e., top-level
definitions), it seems a bit too specific to that particular use case
to use globally (pun intended).