test: tweak Gen.Core.Raw.genName to reuse names more often
Previously we were generating random strings. This had the problem that
it would not pick up bugs due to shadowing etc very frequently (for
example, the one fixed in c86ef8313a5081d9d37d70e2990b84203ca6f885). We
now pick from a small set of fixed names 90% of the time, and generate a
random string the other 10%.