test: tweak Gen.Core.Typed to reuse names
Previously every name generated would be unique. Now we chose one of a
small set, or generate a random string. This choice uses the same
strategy as the previous commit.

However, due to generating well typed terms we need to be careful with
shadowing. Since we mostly generate terms in a top-down fashion, this
mostly is not problematic. The one place we have to be careful is when
generating an annotation -- we verbatim emit a type that may have been
generated somewhere above the current AST node. We need to not capture
any of the free variables of this type. To accomplish this we take the
name generated and add a numeric suffix as necessary to avoid the
relevant names. (Note: this turns out to be a fairly small set of names
to avoid, which is rather far from the previous strategy of "always
generate a completely fresh name".)