fix: Gen.Core.Typed avoid more names
In a5192a864e1e2c44a7315c4ca9efbd1f71731af6 we changed Gen.Core.Typed to
reuse names. However, there was a bug: we would (rather rarely) generate
an ill-sorted term, similar to
  Λa. let f : a -> Bool; f = _ in λa.f (_ : a)
Note that in `_ :: a`, `a` is a term variable but appears within an
annotation!

We need to be slightly more conservative about which names we avoid
reusing. We must not shadow any name that occurs free in the type of
some non-shadowed term variable. One stronger alternative would be
"never shadow any type variable",  but we need not go that far. We are
still much more permissive than "never shadow any name".