refactor applyProgAction: get def from selection
Previously we returned both a Prog and a Maybe GVarName, where the name
was the focused definition for the next action. The relationship between
this name and the Selection which the Prog contains was not clear. We
now only return a Prog, and use the Selection in the places we
previously used the name.

This does change in the semantics, but this is very small change for any
users of the Primer library. Indeed, there is only one place in the
testsuite that needed updating.

The semantic changes are:
- MoveToDef now updates the selection
- RenameDef now selects the renamed definiton both by recording in the
  Prog and for the next action's target
- RenameType, RenameCon, RenameTypeParam, AddCon, SetConFieldType,
  AddConField and RenameModule now do not clear the focus of the next
  action, in line with not changing the Selection