test: munge unused eval_response.json golden file
This golden file has never been used in this repo. In the Vonnegut repo
it was used as a test, but only for the frontend (explaining why it was
copied into this repo and why it was never used). However, it is a bit
outdated: the serialisation changed in
af584b09708f559ce36d720a00432e47e952a47e to not include "unID" and
"unName" wrappers, and Aeson outputs fields in a different order, with
different spacing.

This commit removes the wrappers, and changes the spacing to match.
We will change the order of fields in a different commit, to obtain
easy-to-read diffs.

This diff is probably best viewed with '--ignore-space-change'.