feat!: CreateDef, RenameModule can target any module
Currently they will return an error if they target anything other than
the one editable module. However, we will shortly enable multiple
editable modules, at which point we need to explicitly state what module
they should affect.

BREAKING CHANGE: this change requires a database migration, as it
changes the representation of `Log`. However, since this is just
serialised to json and stored as a blob in the DB, it requires no schema
changes. Since we have no programs we need to preserve, we decided not
to bother with a migration. This means that DBs created before this
commit will not load with a primer containing this commit. This commit
also changes Primer's API.