fix: RenameCon affects case branch labels
When renaming a constructor 'C' to 'D', we must not only rename every
use of it as a constructor, but also where it appears as a branch label.
For example, in the code
  case x of
    C y -> (True, C 1)
there are two usages of 'C' to rename. Previously we only changed the
'C 1' and not the 'C y'.