fix: AddConField robustness
There are two fixes here. Firstly, ensure that updating case branches is
robust when the scrutinee is also updated (due to being an application
of the given constructor). Secondly, handle the case of an application
of a constructor of a parameterised type, i.e. where the spine includes
type applications.