fix: genTypeDefGroup should never refer to a local var
Since it runs in the WT monad, it may happen to be called with some
non-empty local context (which can contain type variables). We should
ignore this since type definitions are global things, and these locals
would be out of scope.

Another way of solving the problem is to change
propertyWTInExtendedLocalGlobalCxt to extend locals and globals in the
other order. (Since the extension functions prepend some setup actions,
the way it was previously written actually extends with locals first, so
that when genTypeDefGroup runs it may have local type variables in the
context).

Whilst changing either one of these functions would be enough on its
own, we implement both fixes for robustness.

We add a property test which (without the other changes in this commit)
very quickly finds an UnknownTypeVariable error due to a type
definition being generated that refers to a local variable.