refactor: typesInExpr is an AffineTraversal
The optic always would target at most one element, but the way we built
it made this fact non-obvious. Let us spell it out with a different
construction.

The fact that it is now explicitly An_AffineTraversal makes the comment
about not recursing clearer (as if it recursed, it would potentially
target many elements).