test: strengthen non-cyclic unification solution condition
Whilst the previous test was true, the unifier actually goes to some
length to only have a "idempotent" solution. This is stronger, as a
solved uv should never appear in the RHS of any unifier. Let's test this
property instead. (This clearly implies that any solution is non-cyclic
in the old sense, as they can only ever non-trivially apply once!)