fix: bound vars shadow unification variables
Previously we would not notice that we had gone under a `forall` with
the same name as a unification variable, and give a "solution" based on
unifying (an occurence of) that bound variable as if it were the
same-named unification variable.

There is a question of why our property tests did not pick up this bug,
in particular hprop_sub_unifies. As far as I can tell, this is because
generating a good test distribution is hard (or at least our current
distribution is unlikely to hit this case). This is because:
- To successfully unify, we need to generate two terms which are
  substantially similar. This is unlikely when we just generate two
  independent terms and try to unify.
- In particular, we need to generate two terms with a forall in the same
  place. (For this example, in the head, but potentially this could
  happen in a subterm)
- For a forall-headed term, we generate these with a frequency of about
  7%
- Thus for two forall-headed terms, we are below 1%.
- To actually tickle this bug, we need further properties, which
  presumably make it vanishingly unlikely.

I did test that hprop_sub_unifies would have failed if it generated
these terms (i.e. when substituting the purported-before-this-commit
unifier, we get inconsistent (i.e. not equal-up-to-holes) terms).

One potential improvement to our tests is to write a new generator
targeting "things which are likely to unify". This could perhaps be
"generate a pair of terms which are similar", or "given an
already-generated term, modify it slightly".