docs: correct an outdated Haddock comment on findNodeByID
This has been incorrect since it was initially committed in this repo.
Sometime before the history of this repo started it was correct, since
at one point we did not have polymorphic functions. We then did not have
LetType, and so the only current evaluation rule for types (inline a
LetType) also did not exist. Unfortunately when this situation changed
the comment was not updated.