fix: findNodeByID cares about shadowing
findNodeByID both focuses on a node and extracts all let (and letrec,
letType) bindings that scope over that subterm. Prior to this commit, it
extracted every such binding, even ones that were shadowed.

Since findNodeByID is only used internally and we do not rely on it to
find redexes, this bug would not have been visible to a user of primer,
unless they made a bad API call. Here, "bad" means a eval-step call on
an ID that is not a redex because of scoping issues. Note that the only
IDs that "should" be used in such a call are those returned by
'redexes', which does check for shadowing.