chore: Rename available action golden test files.
These aren't proper Haskell source files, so let's not use a `.hs`
extension for them. I've arbitrarily chosen the extension `.fragment`
here to indicate that they're Haskell source fragments, rather than
source files per se.

The main reason for this change is that it means we can auto-generate
our HIE file using `implicit-hie` (which I'd already done in a
previous commit) and run `haskell-language-server` with no arguments
from the project root, so that it'll check the whole project in one
go. Prior to this change, HLS would find these golden test files and
choke on them, unless we manually added a stanza to the HIE file
telling HLS to ignore them. Running HLS this way should be useful as a
CI tool check, which I'll add in a subsequent commit.