refactor: take names from DefMap keys, not values
We change
- variablesInScopeExpr
- buildTypingContext
- viewProg
- uniquifyDefName
- unit_rename_module
to look at the keys in the map, rather than the 'defName' of the value.
These are the same, since 'DefMap' maps 'GVarName's to the 'Def'inition
of that global term. Note that 'Def's know their own name currently.

This commit is in preparation for removing names from 'Def's (as they
are redundantly stored in 'Module's), i.e. in a future commit 'Def's
will be changed so they do not know their own name (and thus only make
total sense when stored in a 'Module' or 'Cxt' where they have
names associated).