refactor!: do not store names in Def
We change 'PrimDef' and 'ASTDef' to not store their 'GVarName'. This is
to avoid redundantly storing names in various maps (in modules and in
the typechecker).

This is a large commit, but is mostly mechanical. We sometimes need to
pass around definitions paired with their names, for those places where
we need both. The only notable changes are that
- 'defaultDefs' is simplified since it does not need to take a
  'ModuleName' argument
- Primer.Examples now return '(GVarName, Def)' pairs rather than just
  'Def'
- we remove some checks and tests that names match, since this is
  guaranteed by construction
- there are some minor serialization changes

BREAKING CHANGE: this change requires a database migration, as it
changes the representation of 'ASTDef' and 'PrimDef'.  However, since
this is just serialised to json and stored as a blob in the DB, it
requires no schema changes. Since we have no programs we need to
preserve, we decided not to bother with a migration. This means that DBs
created before this commit will not load with a primer containing this
commit.