refactor: do not store names in Def
We change 'PrimDef' and 'ASTDef' to not store their 'GVarName'. This is
to avoid redundantly storing names in various maps (in modules and in
the typechecker).

This is a large commit, but is mostly mechanical. We sometimes need to
pass around definitions paired with their names, for those places where
we need both. The only notable changes are that
- Primer.Examples now give '(GVarName, Def)' pairs rather than just
  'Def'
- there are some minor serialization changes
- 'defaultDefs' is simplified since it does not need to take a
  'ModuleName' argument
- we remove some tests that names match, since this is guaranteed by
  construction