refactor: take names from TypeDefMap keys, not values
We change
- viewProg
- checkTypeDefs
- Gen.Core.Typed.genChk
- Gen.Core.Typed.genWTType
- unit_defaultFullProg_no_clash
- hprop_con
to look at the keys in the map, rather than the 'typeDefName' of the
value.  These are the same, since 'TypeDefMap' maps 'TyConName's to the
corresponding 'TypeDef'inition. Note that 'TypeDef's know their own name
currently.

This commit is in preparation for removing names from 'TypeDef's (as
they are redundantly stored in 'Module's), i.e. in a future commit
'TypeDef's will be changed so they do not know their own name (and thus
only make total sense when stored in a 'Module' or 'Cxt' where they have
names associated).