refactor!: do not store names in TypeDef
We change 'PrimTypeDef' and 'ASTTypeDef' to not store their 'TyConName'.
This is to avoid redundantly storing names in various maps (in modules
and in the typechecker).

This is a large commit, but is mostly mechanical. We sometimes need to
pass around type definitions paired with their names, for those places
where we need both. The only notable changes are that
- there are some changes to the API of the AddTypeDef action
- we remove some checks and tests that names in maps are consistent,
  since this is guaranteed by construction
- we remove TDIMalformed, as by construction this error case can never
  happen
- there are some minor serialization changes

BREAKING CHANGE: this change requires a database migration, as it
changes the representation of 'ASTTypeDef' and 'PrimTypeDef'. However,
since this is just serialised to json and stored as a blob in the DB, it
requires no schema changes. Since we have no programs we need to
preserve, we decided not to bother with a migration. This means that DBs
created before this commit will not load with a primer containing this
commit.