feat: Don't export `App`'s constructor.
Instead, we export a smart constructor, of sorts. For the moment, this
smart constructor simply wraps the `App` constructor and doesn't
actually do anything "smart," but in a subsequent commit, we'll change
it such that the smart constructor guarantees that the `App` is
correct by construction, which is not something we've been able to
guarantee to this point.