feat!: Remove `InitialApp` support.
We added this feature in:

https://github.com/hackworthltd/vonnegut/commit/1083242b69885c21203591310a6659b6090d76c2

It was never intended for production use; it was simply a
quick-and-dirty way to allow us to create programs that are devoid of
our built-in types, a feature we used in our initial student testing
in 2021. Now that we have module support, our plan is to make it
possible to start with relatively arbitrary initial program
state (e.g., prelude or no prelude, start with another student's
program, etc.), so we remove this antiquated feature in preparation
for that work.

This commit makes a breaking change to the OpenAPI spec, because it
removes the `InitialApp` parameter for the `createSession` verb. This
is of no consequence, because we don't currently have any OpenAPI
clients which depend on this functionality. Of course, this new API
will also eventually be replaced by something that supports variable
initial program state (see above), but for now, as of this commit, one
can only create `newApp`-based `App`s via the OpenAPI `createSession`
verb.

This also breaks `App` database serialization, and would normally
require a database migration. However, we've not bothered with the
migration as we don't currently have any programs we need to preserve.

The exact initial app state *is* still required, because it's the
basis of our current undo system. Therefore, we still need to keep
around a copy of the initial app state, which this commit does.

Side note for posterity: prior to this change, many of our tests
"cheated" by replacing the app's initial program with an arbitrary
program suitable for testing, but there was previously no mechanism
for ensuring that the initial program matched the arbitrary test
program, since we only supported the 2 hard-wired initial programs.
This never caused any problems because the tests that do this don't
test any undo features. In any case, these tests are now "correct" in
the sense that their arbitrary initial programs are properly recorded
as the initial program state in the `App`.