fix: renaming a let could cause an out-of-scope reference
We would rename `λx.let x=x in x` to `λx.let y=y in y`, where the `y` on
the RHS of the binding is an unbound reference. We should not rename in
the bound value of a non-recursive let!

This commit makes it possible to cause shadowing by renaming a let, but
this does not change the semantics of a program (we still forbid
capture: see unit_RenameLet_no_capture_1).