chore: upgrade tasty-discover to 4.2.4
We would like to upgrade tasty-hedgehog to 1.2.0.0, but this introduces
a change (deprecating `testProperty`) which forces large (mechanical)
changes to our codebase, and an upgrade of tasty-discover to stay -Wall
clean. It also needs configuration for hlint, to avoid lint errors. We
separate these three upgrades into their own commits.