doc: Move warning about `Ord` and `Enum` on `ID` to public doc.
This will make it clear to anyone who uses Primer as a library that
they shouldn't rely on these instances.

Note that we've already morally flouted this warning ourselves by
virtue of the fact that a) we increment `ID`s in `MonadFresh ID`
instances in our own production (non-test) code, and b) implicitly
rely on the fact that the maximum `ID` value plus 1 is a valid "next
`ID`". However, as noted in the new Haddock, these implementation
details are more or less not visible to users of the exported Primer
interface. A subsequent commit will make that even more so the case.