feat: Make `mkAppSafe` safer.
Specifically, `mkAppSafe` can now automatically determine the next
valid `ID` of the provided `Prog`.

I've also added additional warnings on `mkApp` to discourage people
from using it. It's tempting to remove it altogether, but I think that
it will still be the preferred way to construct new apps when one has
just built a simple `Prog` from a single module and still has the
valid next `ID` for that single module available. `mkAppSafe` is
helpful for the more general case when one is building a more
complicated `App` from a `Prog` with multiple modules.