test: Flesh out the `NullDbT` implementation.
The `NullDbT` type now properly implements the `MonadDb` spec and
behaves as one would expect from a `MonadDb` implementation, save for
the fact that it's not backed by a persistent store. As before, do not
use this type in production!

In this commit, we also test the `NullDbT` implementation fairly
thoroughly. This may seem a bit odd for a database implementation
that's only intended for testing use itself, but we want to be sure
that our core implementation behaves as expected before building more
tests on that foundation.

There are some serious DRY violations in the test code in this commit.
Much of it is copied from our `primer-rel8` tests. We really need to
factor out a lot of reusable testing code into a separate lib. (See
https://github.com/hackworthltd/primer/issues/273.) However, I'll
leave that work for a later PR, as it will require a pretty
substantial amount of refactoring.