refactor: Add `Control.Monad.Catch` to `Foreword` exports.
We adopt `exceptions`'s way of doing things. It was already
near-universal, anyway. There were a few exceptions (hah), which have
been made to conform in this commit.

(On a somewhat tangential note, previously `primer-service` was not
using `Foreword`, but now it is.)

Fixes https://github.com/hackworthltd/primer/issues/534.