chore: upgrade tasty-hedgehog to 1.2.0
Note that 'testProperty' has been deprecated in favour of
'testPropertyNamed'.

This upstream change was to allow hedgehog and tasty to have different
notions of the test's name: tasty wants a display name and hedgehog
wants a top-level value that can be used with 'recheck' in ghci.

Since (for Tests.Question.test_laws) we don't have such a value, we just
use the static string "<internal-property>". Hopefully this means that
the incorrect hedgehog instructions to 'recheck' any failing properties
under 'test_laws' will be obviously wrong.  Note that the instructions
to rerun via tasty are correct, and are the supported way to run the
tests. This is an improvement on the status quo where the 'recheck'
instructions were less obviously incorrect.