feat: Add new `addSession` and `getApp` API methods.
These API methods are special cases, and we don't expect typical API
clients to use them. `addSession` is intended for pre-seeding a new
Primer database with some useful pre-existing programs, and `getApp`
is useful only for testing.

For now, we've chosen not to add them to the HTTP API, though that
would have some advantages. The decision whether or not to do that is
tracked here:

https://github.com/hackworthltd/primer/issues/550

Note that this commit fixes
https://github.com/hackworthltd/primer/issues/546.