feat: Add a `--seed-db` option to `primer-service`.
When this flag is specified on the command line, `primer-service` will
seed the database with a list of predefined `App`s. This is useful for
getting started with some useful programs in a new database.

It's arguable that we should do this with a separate executable that
would connect to the Primer API and insert the `App`s via the HTTP
API. In addition to a separation of concerns, this would also be a
nice way to test that the Primer HTTP API is functional after starting
`primer-service`. Perhaps in the future we should do that. However, at
the moment, the `addSession` API method is not available via the HTTP
API, because we don't serialize full `App`s to and from JSON -- the
frontend only needs a simpler representation of Primer programs.

Whether we should add these JSON instances is tracked separately here:

https://github.com/hackworthltd/primer/issues/550