feat: Add a `push-docker-image` script.
Note: this script is only useful if you have a GitHub PAT which allows
you to write org-level packages. Its intended use case is for pushing
images that haven't (yet) been committed to the main branch, or in
case CI is down and/or can't push images for some reason.