refactor: Script the `primer-service` Docker container entrypoint.
We replace the `primer-service` Docker `Cmd` with an `Entrypoint` that
runs a shell script. We make this change for the following reasons:

* It will be useful to run some pre-flight operations in the container
before launching `primer-service`. For example, we might want to
ensure that the database schema version matches what `primer-service`
expects, and this is trivial to do with a `sqitch` script.

* `Entrypoint` is more appropriate for our use case than `Cmd` (we are
running a fixed service, not a command-oriented container). For more
details on the rationale, see
https://www.ctl.io/developers/blog/post/dockerfile-entrypoint-vs-cmd/

Note that, for the moment, we don't run any commands in the entrypoint
script except for `primer-service`, so this commit doesn't materially
change the behavior of the container. We are also careful here to use
`exec` in the shell script, so that the shell will replace itself with
the `primer-service` process, ensuring that signals sent to the
container will be handled by `primer-service` and not the shell (once
it has actually exec'ed `primer-service`, of course).