test: property tests that TC+smartholes is idempotent
Adds property tests ensuring that (for well-typed, but not necessarily
smartholes-normalised input) typechecking a definition multiple times
with smartholes is the same as doing so once, with the caveat that bound
names in TypeCache metadata may change each time.

This test would pick up the bugs fixed in commits
  fix: smartholes should add checked-at info when creates a hole
  fix: smartholes should not elide holes with holey-ann contents
This test only passes since it compares expressions up to alpha
in the TypeCache (c.f. the commit
  test: smartholes may alpha convert TypeCache
)

It would be interesting to write the same test but for illtyped
expressions. However this is left as further work since we do not
currently have generators for well-scoped but ill-typed expressions.