fix: smartholes should not elide holes with holey-ann contents
Otherwise it is not idempotent -- we create such things and would change
them the next time the TC runs.

Note that we create such things purposefully: we prefer to have an
explicit hole rather than an annotation of TEmptyHole which acts as a
hole (in the sense that it makes an otherwise ill-typed program type
correct)

However, we can elide a hole-with-trivial-annotation (when it
typechecks): e.g. from {? t : ? ?} we can give t, if it is type correct.
These may happen due to edits of an already-holey program, but we will
not create them directly. This improves the results for the test case
unit_smart_remove_clean_case.