test: Add a `primer-service` Docker image smoke test.
At the moment, all this does is test that the reported API version is
what we expect. This at least ensures that the `primer-service` app
has connected to the database and is serving network requests.

Note that we go to a lot of trouble here to ensure that the Nix
flake's executables are cached for the NixOS tests. For some reason
that I haven't determined after some sleuthing, if we reference the
flake's executables by adding the `primer` Nix overlay to the NixOS
VMs' package set, Nix tries to build not only the `primer-service`
Docker image from scratch, but it also kicks off a GHC 8.10.7
build (presumably in order to bootstrap GHC 9.2.3), rather than using
the cached versions of all of these packages.