fix: Set locale properly for `primer-service` Docker image.
After 5eef6bbde53af2b5585870ef124bbce417f21a71, unless we set the
locale explicitly, the `primer-service` Docker image will fail with
the following error:

```
primer-service: <stdout>: commitAndReleaseBuffer: invalid argument (invalid character)
```

I found the fix here:

https://stackoverflow.com/a/63751678/2109676

Note that my original attempt to fix this issue was to set `LANG` to
`en_US.UTF-8`. However, that did not appear to resolve the issue, so I
resorted to `C.UTF-8` per the Stack Overflow answer. (It's not
particularly important which UTF-8-capable locale we use, as the
banner is probably the only place where we care, though we may need to
revisit this if we ever print Unicode names from students' programs in
the `primer-service` logs.)