refactor: Add "git-" prefix to Primer version number.
Our version numbers are based on the git SHA, so this makes it a bit
more obvious. (Adding a "git-" prefix is also a relatively common
convention.)

Note that due to https://github.com/NixOS/nix/issues/4682, when the
repo is dirty, we use the last-modified date in lieu of the git rev,
and therefore the version in these cases will be "git-{ISO date}",
which is incorrect, but we don't really care as we never publish
Docker images generated from dirty repos.