test: smartholes may alpha convert TypeCache
We add a simple test case showing that smartholes is not idempotent
on-the-nose, but (in this one example) is only up to alpha-equivalence
in the TypeCache attached to each node of an Expr.

This is not ideal, but is difficult to fix without changing how fresh
name generation works. Thankfully it only makes minor changes to
anything observable.