feat: Flesh out `API.Tree`
This augments the datatype enough that we can now render trees which match our designs. And we still won't require the frontend to know anything complex about the semantics of Primer programs.

NB. draft frontend tree rendering code for this new type definition has already been written (with only some minor changes to the definition made since), so we can be confident that it is now rich enough.