feat: Use Servant's `NamedRoutes` for OpenAPI bindings.
This change doesn't alter the functionality of the Primer OpenAPI
bindings; it's just a more ergonomic way to compose Servant routes.
This mechanism is not particularly well documented yet. For details,
see:

https://github.com/haskell-servant/servant/pull/1388
https://github.com/haskell-servant/servant/blob/cdd7c34add589fccd52d59047c1a0a8578a3b5f0/servant/CHANGELOG.md#significant-changes

The `NamedRoutes` mechanism should make it much easier to generate
Servant client bindings, when the time comes. It also allegedly
produces nicer error messages.