doc: Remove method and URL from OpenAPI Haddocks.
The method is redundant, as it's already given in the type signature.

The URL isn't checked in any way, so it's not guaranteed to be
correct. Furthermore, with our new record-based Servant API structure,
the API methods are no longer tied to any particular URL, and it's
probably not worth trying to keep the URL in the Haddock up-to-date in
the event of refactoring or restructuring, given that we're already
automatically generating an OpenAPI JSON spec which includes the
correct URL, anyway.