feat: Remove `API.Tree.label` field
We simplify the API, leaving the exact choice of text up to the client. The text we were using here was uniquely determined by the `style` field anyway.

Given that the annotations are now likely to be used for determining the label text, rather than just colour, we also rename from "style" to "flavor".