fix: checkEverything must check all types before all terms
Previously we checked all definitions simultaneously. If smartholes was
turned on, this could result in edits to the types of top-level things.
These edits would not be picked up when checking any other terms that
referred to them. Instead we check all types of top-level bindings and
potentially edit them and only afterwards we check all terms, making
sure we check against the updated type and in an environment with the
updated types.