refactor: expose checkProgWellFormed
We would like to property test that checkEverything is idempotent (which
would catch the bug fixed in the previous commit). The plan is to do so
by generating whole 'Prog's and ensuring that the new
'tcWholeProgWithImports' is idempotent (running with smartholes). We
want to ensure that any generated program is well formed (via the new
'checkProgWellFormed'), so this commit is preparatory work. (Note that
the property we will test is essentially whether smartholes are
idempotent, at the level of an entire program, rather than one
expression. Note also that we only care about initial states which are
valid (but not "smartholes-normalised") programs.)