feat!: Switch `PrimerJSON` to default Aeson encodings.
We could drop `PrimerJSON` entirely here, but we keep it a) to get the
more efficient `toEncoding` implementation generated by
`deriving-aeson` (as we understand it), and b) to provide an escape
hatch if we need to add custom encodings in the future.

BREAKING CHANGE: this changes how all-nullary sum types are serialized
to/from JSON. They are now serialized as plain strings. This
serialization appears to be compatible with `openapi3`'s default
encodings, and should be ergonomically represented in TypeScript, so
this is a beneficial change.