chore(API): Remove Haddocks, replace with `Description`s.
At least for OpenAPI endpoints, these descriptions will be exported in
the API spec, which is where they're most useful.

Note that we also drop the `GET /foo/bar`-style comments here, as
those URLs are a) not checkable and b) can change depending on how the
Servant bits are glued together in Haskell.