refactor: split a testing library
We split hedgehog generators, and some tasty glue into its own library
(as part of the Primer package).  We split these out to enable reuse of
generators between primer and primer-service in the future.  Note that
whilst internal libraries are supported with cabal-version: 2.4, the
ability to set visibility is only available from 3.0.