fix: GlobalNames ToJSON/ToSchema consistency
Note that the old instance (via `Name`) was nonsense: `GlobalName`s and
`Name`s are not serialised the same (one is a record, the other is a
string)! The instance compiled even though `GlobalName` is not
`Coercible` with `Name` since GHC only requires that the corresponding
methods of the two instances are `Coercible. Since `ToSchema a` only
uses `a` via `Proxy`, this is true for any choices of `a`.